Introduce manifest variants
This change updates the config and Makefile to produce multiple variants
of the HNC manifests. The initial list is "default" (same as today),
"default-cm" (same, but configured for cert-manager instead of the
internal cert-controller), and ha (replicated webhooks, using the
internal cert controller).

Also made some updates to the Prow tests to allow for variant testing,
as well as upgrading the ability to test the Prow tests locally.

Tested: verified that default.yaml is the same as the prior
hnc-manager.yaml, apart from some reordered command line args. Verified
that smoke tests work on ha.yaml. Did not do _any_ testing on the
cert-manager variants. Locally ran the prow e2e tests on both the HA and
default configs.